Automatic_V3_for_HEVC
When HEVC stream is detected, tsMuxer automatically changes BD version to V3.
Note that V3 manual selection is still needed for AVC. However it may be more appropriate to include the option as a check box in the "Blu-ray" tab, in lieu of the two radio buttons (UHD Blu-ray ISO, UHD Blu-ray folder) on the main window.